requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Faker>=19.0